Created InputParser.x10 to allow us to read input
The input is in the form of a text file.  The first
three lines contain only a single integer.  Line 1
is the size of the data in the x direction, line 2
in the y direction, and line 3 in the z direction.
The remaining lines of the file are comma-separated
lists of Double and should contain exactly as many
entries as the number on line 1.  The number of rows
of data should be exactly equal to y * z.