Added OutputPrinter.x10
This files is meant to take the Arrays used in thermal
to compute the heat flow and print them out in a format
that humans can understand, and that can be read into
MATLAB for an easy visualization.

Four print methods (each called printm) have been defined,
each prints the exact same material but in different ways.
See the .x10 file for details.