Removed old code from thermal in borderFill
The previous comment said that parallelizing the borderFill
method would be tricky, but after I committed I realized that
the edges and corners of the outer cube don't ever affect the
values in the interior cube, so we don't even have to set those
if we don't want to.  Parallelization then becomes trivial.